feat: refactor bot handling and add socket connection with worker management
